phase-1/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
anthropic
python-dotenv
pydantic
python-multipart
